async_logger_orig: create unit tests that fail
Use googletest's TEST_P feature in order to create unit tests that
take input from stdin and write to an ostringstream.  The result is
unit tests that fail without modification to the original code.

The tests pass about 2/3 of the time.  If the logger thread fails to
produce the output from the test1 thread, then the test itself exits
normally.  If the test completes before the detached logger thread has
finished its work, then attempts to access the queue_ result in
use-after-free since the test's dtor will have deleted it.