async_logger: create new unit tests for improved code from Parvu
Adding the synchronization and atomicity features suggested by Parvu
makes the application from the ACCU {cvu} article always pass.  The
code from the application's main() function runs with as few changes
as possible in the constructor and destructor of the test class.

Two main changes were necessary for the test. The first is redirection
of stdout to an ostringstream so that the test can compare the actual
output to what is expected.  The second is to add an additional lock
in order to prevent the comparison from running before the logger
thread terminates.  The logger thread only terminates when the
async_logger object destructor runs, and that does not happen until
the test class itself is destroyed.   The solution is to put the
actual test statements in the TearDown() method of the test class
since, per $HOME/gitsrc/googletest/googletest/docs/faq.html,

    For each TEST_F, googletest will create a fresh test fixture
    object, immediately call SetUp(), run the test body, call
    TearDown(), and then delete the test fixture object.

The result is a test that passes 1000 times in a row, while the
original code only passed two or three times in a row with a large
argc.