arg_classifier: determine how libgcc classifies C++ types
Provide a function that returns GCC's type_class for a parameter.

The FOLLY_SDT macro defined in the StaticTracepoint.h header is
implemented for x86 with the __builtin_classify_type() function in
StaticTracepoint-ELFx86.h.  The logic there in the
FOLLY_SDT_IS_ARRAY_POINTER macro defines parameters in GCC's
pointer_type_class and array_type_class to have sizeof(void*).  A
problem arises when C++ std::string parameters are passed to the
FOLLY_SDT macro, as GCC classifies them as record_type_class with size
32 bytes.