USDT: control tracepoint behavior with semaphore
Insert the tracing endpoint probe with a semaphore and then only
perform the time-consuming integration if a tracer is active.  The
boolean test always returned false when only one of the two
operation_start probes was declared with a semaphore.  "readelf -n"
shows the presence of the semaphore in async_logger_improved .notes
section:

Displaying notes found in: .note.stapsdt
  Owner                Data size        Description
  stapsdt              0x00000056       NT_STAPSDT (SystemTap probe descriptors)
    Provider: async_logger_improved
    Name: operation_end
    Location: 0x00000000004cd5aa, Base: 0x0000000000000000, Semaphore: 0x000000000053d140
    Arguments: -8@120(%rsp) -8@112(%rsp)