arg_classifier: hide the template type via erasure
Following http://davekilian.com/cpp-type-erasure.html and
https://www.cplusplus.com/articles/oz18T05o/ , create an opaque
wrapper for PODs and complex data types.  Establish that the wrapped
objects of various types can be stored together in a container.

As established in the cited examples, retrieving the underlying type
from the wrapped objects is simple.  Peeking at the objects themselves
will require teaching the wrapped-object base class how to deduce the
object type.  The public-API functions must call the base-class
functions since the whole point of wrapping is that the underlying
type is "erased" at the container level.