bpftrace: print longest contention times of locks
Exercise 3.2.   Apparently some the printing of some maps can only be suppressed
by clear()-ing them, as they cannot be deleted.   My guess is that's because the
histogram, which should be printed, depends on them.

Another oddity is that printing out the maps indexed by tid shows tid 0.  While
lock address of 0 is clearly a bug, tid appears to be the core kernel, since "ps
-eLf" reports a PPID of 0 for systemd and kthreadd.  Thus perhaps tid 0 is the
core kernel.